refactor(libharness): share the producer-grade composition; split lint-ceiling helpers
runProducersAndGrade is the one composition the runner and the grade
subcommand share (the design's contract); trace-split.js and the report
helpers are behavior-free extractions that keep runner.js and report.js
under the file-length and complexity ceilings.
