+++ libraries/libharness/src/benchmark/runner.js
@@ -19,7 +19,6 @@
  * stream.
  */
 
-import { createInterface } from "node:readline";
 import { join, resolve as resolvePath } from "node:path";
 
 import { DEFAULT_ENV_ALLOWLIST, createRedactor } from "../redaction.js";
@@ -31,8 +30,9 @@ import { runJudge } from "./judge.js";
 import { validateResultRecord } from "./result.js";
 import { runInvariants } from "./invariants.js";
 import { runHiddenTests } from "./hidden-tests.js";
-import { gradeChecks, mergeRows, normalizeGrade } from "./grade.js";
+import { runProducersAndGrade } from "./grade.js";
 import { assertJudgeProfileStaged, loadTaskFamily } from "./task-family.js";
+import { splitAndSummarize } from "./trace-split.js";
 import { createWorkdirManager } from "./workdir.js";
 import { CellScheduler } from "./scheduler.js";
 
@@ -303,35 +303,13 @@ export class BenchmarkRunner {
       const breakdown = agentRun.costBreakdown ?? { agent: 0, supervisor: 0 };
       const graded = await this.#gradeCell(family, task, workdir);
       const { invariants, hiddenRows, engineError, rows, grade } = graded;
-      let judgeVerdict = null;
-      let judgeCost = 0;
-      if (task.paths.judge) {
-        const judgeContext = await this.#buildJudgeContext(
-          task,
-          workdir,
-          skillSetHash,
-        );
-        const judgeResult = await this._runJudgeHook(
-          task,
-          workdir,
-          { ...grade, rows },
-          {
-            query: this.query,
-            model: this.judgeModel,
-            judgeProfile: this.profiles.judge ?? undefined,
-            profilesDir: judgeProfilesDir,
-            runtime: this.runtime,
-          },
-          judgeContext,
-        );
-        judgeCost = judgeResult.costUsd ?? 0;
-        // The record's judgeVerdict carries only the verdict + summary; the
-        // judge's cost is folded into costUsd / costBreakdown instead.
-        judgeVerdict = {
-          verdict: judgeResult.verdict,
-          summary: judgeResult.summary,
-        };
-      }
+      const { judgeVerdict, judgeCost } = await this.#judgeCell({
+        task,
+        workdir,
+        gradeResult: { ...grade, rows },
+        skillSetHash,
+        judgeProfilesDir,
+      });
       const judgePass =
         judgeVerdict === null || judgeVerdict.verdict === "pass";
       const verdict = grade.verdict === "pass" && judgePass ? "pass" : "fail";
@@ -386,33 +364,62 @@ export class BenchmarkRunner {
   }
 
   /**
-   * Run both check-row producers against the post-run CWD and grade the
-   * merged rows. An engine throw is grader fault: its message lands on
-   * `hiddenTests.error` and health fails, so a crashed grader can never mint
-   * marks from rows it happened to emit first. Restoration happens inside
-   * the engine, so the judge (which runs after) sees the workdir exactly as
-   * the agent left it.
+   * Run the judge (when the task ships a template) over the grade result.
+   * The record's judgeVerdict carries only the verdict + summary; the
+   * judge's cost is folded into costUsd / costBreakdown instead.
    */
-  async #gradeCell(family, task, workdir) {
+  async #judgeCell({
+    task,
+    workdir,
+    gradeResult,
+    skillSetHash,
+    judgeProfilesDir,
+  }) {
+    if (!task.paths.judge) return { judgeVerdict: null, judgeCost: 0 };
+    const judgeContext = await this.#buildJudgeContext(
+      task,
+      workdir,
+      skillSetHash,
+    );
+    const judgeResult = await this._runJudgeHook(
+      task,
+      workdir,
+      gradeResult,
+      {
+        query: this.query,
+        model: this.judgeModel,
+        judgeProfile: this.profiles.judge ?? undefined,
+        profilesDir: judgeProfilesDir,
+        runtime: this.runtime,
+      },
+      judgeContext,
+    );
+    return {
+      judgeVerdict: {
+        verdict: judgeResult.verdict,
+        summary: judgeResult.summary,
+      },
+      judgeCost: judgeResult.costUsd ?? 0,
+    };
+  }
+
+  /**
+   * Run both check-row producers against the post-run CWD and grade the
+   * merged rows via the shared derivation. Restoration happens inside the
+   * engine, so the judge (which runs after) sees the workdir exactly as the
+   * agent left it.
+   */
+  #gradeCell(family, task, workdir) {
     const ctx = {
       cwd: workdir.cwd,
       port: workdir.port,
       runDir: workdir.runDir,
       familyDir: family.rootPath,
     };
-    const invariants = await this._runInvariantsHook(task, ctx, this.runtime);
-    let hiddenRows = [];
-    let engineError = null;
-    try {
-      const hidden = await this._runHiddenTestsHook(task, ctx, this.runtime);
-      hiddenRows = hidden.details;
-    } catch (e) {
-      engineError = e;
-    }
-    const rows = mergeRows(invariants.details, hiddenRows);
-    const healthy = invariants.exitCode === 0 && !engineError;
-    const grade = normalizeGrade(gradeChecks(rows, healthy));
-    return { invariants, hiddenRows, engineError, rows, healthy, grade };
+    return runProducersAndGrade(task, ctx, this.runtime, {
+      runInvariants: this._runInvariantsHook,
+      runHiddenTests: this._runHiddenTestsHook,
+    });
   }
 
   /**
@@ -656,67 +663,6 @@ async function writeRecord(stream, record) {
   await new Promise((res, rej) => {
     stream.write(line, (err) => (err ? rej(err) : res()));
   });
-}
-
-/**
- * Split the combined supervisor trace into agent and supervisor files and
- * extract turn count and submission in a single pass. Agent-source events go
- * to `agentPath`; supervisor and orchestrator events go to `supervisorPath`.
- *
- * Cost is deliberately not summed here — the caller derives it from the same
- * combined trace via `sumTraceCost`, so there is one cost path across the
- * benchmark, callback, and `fit-trace cost` consumers.
- */
-// biome-ignore lint/complexity/noExcessiveCognitiveComplexity: stream-splitting state machine
-async function splitAndSummarize(
-  runtime,
-  combinedPath,
-  agentPath,
-  supervisorPath,
-) {
-  const fs = runtime.fs;
-  const agentStream = fs.createWriteStream(agentPath);
-  const supStream = fs.createWriteStream(supervisorPath);
-  const rl = createInterface({
-    input: fs.createReadStream(combinedPath),
-    crlfDelay: Infinity,
-  });
-  let turns = 0;
-  let submission = "";
-  for await (const line of rl) {
-    if (!line.trim()) continue;
-    let event;
-    try {
-      event = JSON.parse(line);
-    } catch {
-      continue;
-    }
-    const target = event.source === "agent" ? agentStream : supStream;
-    target.write(line + "\n");
-    const inner = event.event;
-    if (!inner) continue;
-    if (event.source === "agent" && inner.type === "assistant") {
-      const text = extractText(inner);
-      if (text) submission = text;
-    }
-    if (event.source === "orchestrator" && inner.type === "summary") {
-      turns = inner.turns ?? 0;
-    }
-  }
-  await Promise.all([
-    new Promise((r) => agentStream.end(r)),
-    new Promise((r) => supStream.end(r)),
-  ]);
-  return { turns, submission };
-}
-
-function extractText(inner) {
-  const content = inner.message?.content ?? inner.content;
-  if (!Array.isArray(content)) return null;
-  for (let i = content.length - 1; i >= 0; i--) {
-    if (content[i].type === "text" && content[i].text) return content[i].text;
-  }
-  return null;
 }
 
 /**
